refactor: rename Z21Service to Z21CommandService and update related imports
