+++ application-platform.iml
@@ -18,6 +18,8 @@
       <sourceFolder url="file://$MODULE_DIR$/libs/z21-suite/domain/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/libs/z21-suite/protocol/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/libs/z21-suite/z21/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/libs/z21-suite/server-utils/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/libs/z21-suite/shared/src" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/.angular" />
       <excludeFolder url="file://$MODULE_DIR$/.github" />
       <excludeFolder url="file://$MODULE_DIR$/.husky" />
