+++ apps/z21-server/src/infra/ws/app-websocket-server.ts
@@ -32,7 +32,7 @@ export class AppWsServer {
 	 */
 	constructor(
 		private readonly wsServer: WsServer,
-		private logger: Logger
+		private readonly logger: Logger
 	) {}
 
 	/**
