+++ libs/z21-suite/shared/src/index.ts
@@ -3,11 +3,11 @@
  * All rights reserved.
  */
 
-export * from './lib/lan-x-command';
-export * from './lib/lan-x-types';
+export * from './lib/interfaces/lan-x-command';
 export * from './lib/logging/logger';
-export * from './lib/pom-options';
-export * from './lib/shared-types';
-export * from './lib/x-bus-cmd';
-export * from './lib/x-header';
-export * from './lib/z21-lan-header';
+export * from './lib/types/lan-x-types';
+export * from './lib/types/pom-options';
+export * from './lib/types/shared-types';
+export * from './lib/types/x-bus-cmd';
+export * from './lib/types/x-header';
+export * from './lib/types/z21-lan-header';
