+++ libs/z21-suite/shared/src/lib/types/z21-lan-header.ts
