+++ libs/z21-suite/z21/src/index.ts
@@ -12,7 +12,7 @@ export * from './lib/helper/x-bus-encoder';
 export * from './lib/lanx/decode/decoder';
 export * from './lib/lanx/dispatch';
 export * from './lib/lanx/encoder';
-export * from './lib/services/z21-service';
+export * from './lib/services/z21-command-service';
 export * from './lib/system/decode-system-state';
 export * from './lib/system/derive-track-flags';
 export * from './lib/udp/udp-types';
